deps: update V8 to 12.7.137

+++ deps/v8/src/interpreter/bytecode-array-writer.h
@@ -58,7 +58,7 @@ class V8_EXPORT_PRIVATE BytecodeArrayWriter final {
   template <typename IsolateT>
   EXPORT_TEMPLATE_DECLARE(V8_EXPORT_PRIVATE)
   Handle<BytecodeArray> ToBytecodeArray(IsolateT* isolate, int register_count,
-                                        int parameter_count,
+                                        uint16_t parameter_count,
                                         Handle<TrustedByteArray> handler_table);
 
   template <typename IsolateT>
